If .debug_info section not present, exit with 0
There are two errors switched off by --quiet (provided we don't run with -o):
- DWARF compression not beneficial
- .debug_info section not present
If we run into the first error, we exit with '0', but if we run into the
second error, we exit with '1'.

It seems that if the error is unimportant enough to not print with --quiet,
it's not worth exiting with '1' either.  Fix this.

2019-07-11  Tom de Vries  <tdevries@suse.de>

	PR dwz/24766
	* dwz.c (read_dwarf): Return 3 if .debug_info not present.
	(dwz): If read_dwarf returned 3, exit with 0.
	* testsuite/dwz.tests/objcopy-strip-debug.sh: Update.

+++ testsuite/dwz.tests/objcopy-strip-debug.sh
@@ -11,8 +11,20 @@ if ! grep -q "\.debug_info section not present" dwz.err; then
     exit 1
 fi
 
-[ $status -eq 1 ]
+[ $status -eq 0 ]
 
 cmp 1 1.saved
+
+if dwz 1 -o 2 2>dwz.err; status=$?; then
+    true
+fi
+
+if ! grep -q "\.debug_info section not present" dwz.err; then
+    cat dwz.err
+    exit 1
+fi
+
+[ $status -eq 1 ]
+[ ! -f 2 ]
 
 rm -f 1 1.saved dwz.err
